backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
mysql-connector-python==9.1.0
python-dotenv==1.1.0
